Move code from FormatClientService into JavaCodeAssistClient
