+++ plugin-java/che-plugin-java-ext-jdt/che-jdt-ext-machine/src/main/java/org/eclipse/jdt/internal/corext/format/CheCodeFormatterInitializer.java
@@ -18,7 +18,6 @@ import java.util.Map;
 /**
  * @author Roman Nikitenko
  */
-
 public class CheCodeFormatterInitializer extends AbstractPreferenceInitializer {
 
     @Override
