+++ plugin-java/che-plugin-java-ext-jdt/che-jdt-ext-machine/src/main/java/org/eclipse/jdt/internal/corext/format/CheCodeFormatterOptions.java
@@ -24,7 +24,6 @@ import java.util.Map;
 /**
  * @author Roman Nikitenko
  */
-
 public class CheCodeFormatterOptions {
     private static final Logger LOG               = LoggerFactory.getLogger(CheCodeFormatterInitializer.class);
     private static final String DEFAULT_CODESTYLE = "che-codestyle-eclipse_.xml";
