+++ plugin-java/che-plugin-java-ext-jdt/che-jdt-ext-machine/src/main/java/org/eclipse/jdt/internal/corext/format/Formatter.java
@@ -23,7 +23,6 @@ import java.util.List;
 /**
  * @author Roman Nikitenko
  */
-
 public class Formatter {
 
     /**
